pyproject.toml
[project]
name = "synapse-archive-bot"
version = "0.1.0"
description = "Archive bot to all messages in a Synapse server"
readme = "README.md"
requires-python = ">=3.10.0,<4.0.0"
dependencies = [
    "matrix-nio>=0.25.2",
    "psycopg[binary]>=3.3.3",
]

[dependency-groups]
dev = [
    "matrix-synapse>=1.148.0",
    "pyright>=1.1.408",
    "ruff>=0.15.2",
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py310"
